Teljes project src dir refactor.
Packagek lértehozása minden "szubkategóriának".
