+++ src/main/java/com/dino/tecton/SingleHyphaTecton.java
@@ -1,4 +1,6 @@
-package com.dino;
+package com.dino.tecton;
+
+import com.dino.core.Hypha;
 
 /**
  * Egy olyan konkrét, példányosítható tektontípus, amin egyszerre csak egy gombafonál lehet jelen.
@@ -19,7 +21,7 @@ public class SingleHyphaTecton extends Tecton {
      * @param h A kezelendő gombafonál
      */
     @Override
-    protected void handleHypha(Hypha h) {
+    public void handleHypha(Hypha h) {
         //TODO
     }
 }
